+++ packages/apollo-server-core/src/ApolloServer.ts
@@ -267,6 +267,20 @@ export class ApolloServerBase<
 
     if (!requestOptions.cache) {
       requestOptions.cache = new UnboundedCache();
+
+      if (
+        !isDev &&
+        (requestOptions.persistedQueries === undefined ||
+          (requestOptions.persistedQueries &&
+            !requestOptions.persistedQueries.cache))
+      ) {
+        this.logger.warn(
+          'Persisted queries are enabled and are using an unbounded cache. Your server' +
+            ' is vulnerable to denial of service attacks via memory exhaustion. ' +
+            'Set `cache: "bounded"` or `persistedQueries: false` in your ApolloServer ' +
+            'constructor, or see FIXME:DOCS for other alternatives.',
+        );
+      }
     }
 
     if (requestOptions.persistedQueries !== false) {